# docs/api/swagger.yaml
definitions:
  APIErrorResponse:
    properties:
      code:
        type: integer
      details: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  internal_modules_project_handler.DeleteResponse:
    properties:
      status:
        type: string
    type: object
  internal_modules_project_handler.EnvCreateRequest:
    properties:
      is_base64:
        example: false
        type: boolean
      key:
        example: API_KEY
        type: string
      path:
        example: android/app/google-services.json
        type: string
      project_id:
        example: 2
        type: integer
      type:
        description: '"env" or "file"'
        example: env
        type: string
      value:
        example: super-secret
        type: string
    type: object
  internal_modules_project_handler.EnvDTO:
    properties:
      created_at:
        type: string
      id:
        example: 1
        type: integer
      is_base64:
        example: false
        type: boolean
      key:
        example: API_KEY
        type: string
      path:
        example: android/app/google-services.json
        type: string
      project_id:
        example: 2
        type: integer
      type:
        description: '"env" or "file"'
        example: env
        type: string
      updated_at:
        type: string
      user_id:
        example: 1
        type: integer
      value:
        example: super-secret
        type: string
    type: object
  internal_modules_project_handler.EnvListResponse:
    properties:
      envs:
        items:
          $ref: '#/definitions/internal_modules_project_handler.EnvDTO'
        type: array
    type: object
  internal_modules_project_handler.EnvResponse:
    properties:
      env:
        $ref: '#/definitions/internal_modules_project_handler.EnvDTO'
    type: object
  internal_modules_project_handler.EnvUpdateRequest:
    properties:
      is_base64:
        example: false
        type: boolean
      key:
        example: API_KEY
        type: string
      path:
        example: android/app/google-services.json
        type: string
      project_id:
        example: 2
        type: integer
      type:
        description: '"env" or "file"'
        example: env
        type: string
      value:
        example: super-secret
        type: string
    type: object
  internal_modules_project_handler.FlutterVersion:
    properties:
      channel:
        type: string
      version:
        type: string
    type: object
  internal_modules_project_handler.FlutterVersionsResponse:
    properties:
      versions:
        items:
          $ref: '#/definitions/internal_modules_project_handler.FlutterVersion'
        type: array
    type: object
...
info:
  contact:
    email: support@flotio.dev
    name: Flotio Team
  description: Flotio Core API — authentication, environment/keystore/Google Play
    credentials assets, projects, build pipeline, releases and GitHub integration
    for the Flotio mobile CI platform.
  license:
    name: Proprietary
  title: Flotio Core API
  version: 1.0.0
paths:
...
  /env:
    get:
      consumes:
      - application/json
      description: Get all environment assets/files for the authenticated user, optionally
        filtered by project_id
      operationId: EnvGetHandler
      parameters:
      - description: Filter by Project ID
        in: query
        name: project_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_project_handler.EnvListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get environment assets
      tags:
      - environment
    post:
      consumes:
      - application/json
      description: Create a new environment variable or configuration file for the
        user
      operationId: EnvPostHandler
      parameters:
      - description: Environment asset data
        in: body
        name: env
        required: true
        schema:
          $ref: '#/definitions/internal_modules_project_handler.EnvCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_modules_project_handler.EnvResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Create environment asset
      tags:
      - environment
  /env/{envId}:
    delete:
      consumes:
      - application/json
      description: Delete an environment asset by its ID
      operationId: EnvDeleteByIdHandler
      parameters:
      - description: Environment asset ID
        format: int64
        in: path
        name: envId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_project_handler.DeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete environment asset
      tags:
      - environment
    get:
      consumes:
      - application/json
      description: Get a specific environment asset by its ID
      operationId: EnvGetByIdHandler
      parameters:
      - description: Environment asset ID
        format: int64
        in: path
        name: envId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_project_handler.EnvResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get environment asset by ID
      tags:
      - environment
    put:
      consumes:
      - application/json
      description: Update an existing environment asset
      operationId: EnvPutByIdHandler
      parameters:
      - description: Environment asset ID
        format: int64
        in: path
        name: envId
        required: true
        type: integer
      - description: Updated environment asset data
        in: body
        name: env
        required: true
        schema:
          $ref: '#/definitions/internal_modules_project_handler.EnvUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_project_handler.EnvResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Update environment asset
      tags:
      - environment
  /envs:
    get:
      consumes:
      - application/json
      description: 'Alias of GET /env: get all environment assets/files for the authenticated
        user, optionally filtered by project_id'
      operationId: EnvsGetHandler
      parameters:
      - description: Filter by Project ID
        in: query
        name: project_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_project_handler.EnvListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get environment assets
      tags:
      - environment
  /flutter/versions:
    get:
      consumes:
      - application/json
      description: Get a list of available Flutter versions from the official releases
      operationId: VersionsGetHandler
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_project_handler.FlutterVersionsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get available Flutter versions
      tags:
      - flutter
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"